Modified spatial filters to use ol.layer.Layer
Modified the spatial filter functionality to support filtering on
multiple features and to store the filters in a separate, editable
layer. Filters can now be edited, modified, and deleted using the
built-in feature editing controls. Previously filters could only be
added and filtering could only be performed on a single feature.