Using a radius for "point" spatial filters
Modified the spatial filtering for points to create a circular polygon
of a desired radius around the point and to use that polygon for
filtering.

Previously the point's coordinates were used for the spatial filter. The
previous approach resulted in confusing behavior. For example, selecting
a point, adding it to the spatial filter, and then viewing the filtered
results for the layer would not display the point being filtered on.