MapLoom version is now available in the UI.
It accessed by hovering over the "info" icon in the save map
dialog. There are not many default dialogs and this was a
predictable landing spot.  The icon itself is small and out of the
way.

The version is added into "scope" by using the window variable
MAPLOOM_VERSION, which is an object containing the version string
and the build date.