Fix for datepicker throwing error.
The date-picker was throwing an error if updateDate() was called
before the scope was able to set the date.  This was causing
an error dialog to be shown to the user when looking at the date panel.

Removing the extra call to "updateDate" and using  dateObjectChanged
has fixed the problem.