Fix for history zooming.
The 'zoom to' feature was not accoutning for different projections.

In most examples, the map is in WebMercator and the layer is sourced
from a WGS84/EPSG:4326 dataset.  This adds projection sensing support
for the layer.  It defaults to the layer's data being in WGS84 if for
some reason a valid CRS cannot be found.