Added a spinner + help prompt for searching.
The previous search dialog would show the user that they could
not find any layers, even before any search was executed. This caused
some confusion as one of the filters must be selected *before*
a search will execute.  This new version of the dialog attempts
to make that more clear by:
1. Showing a 'spinner' while the user is searching.
2. Adding first-time help test to the layers list in order to
   instruct the user to use the options at left to start a search.