Fixed attribute editing error
This error is related to using GetCaps 1.1.1. The projection
was not parsed correctly as the Openlayers GetCaps parser assumes
the 1.3.0 nomenclature.  Therefore, the layer's SRS was not transformed
or interpreted as expected (that is as the CRS as the 1.3.0 spec calls for).

This fix reverts the change to the 1.1.1 version of GetCaps but ultimately
still fixes the zoom-to errors that were happening before.

refs: BEX-469, BEX-418, PR #114